dstu3 measure processor enhancements
