--- cqf-fhir-cr/src/main/java/org/opencds/cqf/fhir/cr/measure/dstu3/Dstu3MeasureEvaluation.java
@@ -1,32 +1,0 @@
-package org.opencds.cqf.fhir.cr.measure.dstu3;
-
-import org.hl7.elm.r1.VersionedIdentifier;
-import org.hl7.fhir.dstu3.model.DomainResource;
-import org.hl7.fhir.dstu3.model.Measure;
-import org.hl7.fhir.dstu3.model.MeasureReport;
-import org.opencds.cqf.cql.engine.execution.CqlEngine;
-import org.opencds.cqf.fhir.cql.LibraryEngine;
-import org.opencds.cqf.fhir.cr.measure.common.BaseMeasureEvaluation;
-
-/**
- * Implementation of MeasureEvaluation on top of HAPI FHIR DSTU3 structures.
- */
-public class Dstu3MeasureEvaluation extends BaseMeasureEvaluation<Measure, MeasureReport, DomainResource> {
-
-    public Dstu3MeasureEvaluation(
-            CqlEngine context, Measure measure, LibraryEngine libraryEngine, VersionedIdentifier versionIdentifier) {
-        super(
-                context,
-                measure,
-                new Dstu3MeasureDefBuilder(),
-                new Dstu3MeasureReportBuilder(),
-                libraryEngine,
-                versionIdentifier,
-                new Dstu3PopulationBasisValidator());
-    }
-
-    @Override
-    protected String getMeasureUrl() {
-        return measure.getUrl();
-    }
-}
