+++ cqf-fhir-cr/src/main/java/org/opencds/cqf/fhir/cr/measure/dstu3/Dstu3MeasureProcessor.java
@@ -1,6 +1,7 @@
 package org.opencds.cqf.fhir.cr.measure.dstu3;
 
 import ca.uhn.fhir.rest.server.exceptions.InvalidRequestException;
+import java.time.ZonedDateTime;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
@@ -18,6 +19,7 @@ import org.hl7.fhir.dstu3.model.MeasureReport;
 import org.hl7.fhir.dstu3.model.Parameters;
 import org.hl7.fhir.instance.model.api.IBaseBundle;
 import org.hl7.fhir.instance.model.api.IPrimitiveType;
+import org.opencds.cqf.cql.engine.execution.CqlEngine;
 import org.opencds.cqf.cql.engine.fhir.model.Dstu3FhirModelResolver;
 import org.opencds.cqf.cql.engine.runtime.Interval;
 import org.opencds.cqf.fhir.api.Repository;
@@ -25,9 +27,11 @@ import org.opencds.cqf.fhir.cql.Engines;
 import org.opencds.cqf.fhir.cql.LibraryEngine;
 import org.opencds.cqf.fhir.cr.measure.MeasureEvaluationOptions;
 import org.opencds.cqf.fhir.cr.measure.common.MeasureEvalType;
+import org.opencds.cqf.fhir.cr.measure.common.MeasureProcessorUtils;
 import org.opencds.cqf.fhir.cr.measure.common.MeasureReportType;
 import org.opencds.cqf.fhir.cr.measure.common.SubjectProvider;
 import org.opencds.cqf.fhir.cr.measure.helper.DateHelper;
+import org.opencds.cqf.fhir.cr.measure.r4.R4PopulationBasisValidator;
 import org.opencds.cqf.fhir.utility.repository.FederatedRepository;
 import org.opencds.cqf.fhir.utility.repository.InMemoryFhirRepository;
 
@@ -35,6 +39,7 @@ public class Dstu3MeasureProcessor {
     private final Repository repository;
     private final MeasureEvaluationOptions measureEvaluationOptions;
     private final SubjectProvider subjectProvider;
+    private final MeasureProcessorUtils measureProcessorUtils = new MeasureProcessorUtils();
 
     public Dstu3MeasureProcessor(Repository repository, MeasureEvaluationOptions measureEvaluationOptions) {
         this(repository, measureEvaluationOptions, new Dstu3RepositorySubjectProvider());
@@ -73,69 +78,42 @@ public class Dstu3MeasureProcessor {
             IBaseBundle additionalData,
             Parameters parameters) {
 
-        if (!measure.hasLibrary()) {
-            throw new InvalidRequestException(
-                    String.format("Measure %s does not have a primary library specified", measure.getUrl()));
-        }
+        checkMeasureLibrary(measure);
 
-        Interval measurementPeriod = null;
-        if (StringUtils.isNotBlank(periodStart) && StringUtils.isNotBlank(periodEnd)) {
-            measurementPeriod = this.buildMeasurementPeriod(periodStart, periodEnd);
-        }
+        Interval measurementPeriod = measureProcessorUtils.buildMeasurementPeriod(periodStart, periodEnd);
 
-        var reference = measure.getLibrary().get(0);
-
-        var library = this.repository.read(Library.class, reference.getReferenceElement());
-
-        var id = new VersionedIdentifier().withId(library.getName()).withVersion(library.getVersion());
-        var context = Engines.forRepository(
-                this.repository, this.measureEvaluationOptions.getEvaluationSettings(), additionalData);
-
-        CompiledLibrary lib;
-        try {
-            lib = context.getEnvironment().getLibraryManager().resolveLibrary(id);
-        } catch (CqlIncludeException e) {
-            throw new IllegalStateException(
-                    String.format(
-                            "Unable to load CQL/ELM for library: %s. Verify that the Library has CQL/ELM content embedded.",
-                            id.getId()),
-                    e);
-        }
-
-        context.getState().init(lib.getLibrary());
-
-        if (parameters != null) {
-            Map<String, Object> paramMap = resolveParameterMap(parameters);
-            context.getState().setParameters(lib.getLibrary(), paramMap);
-            // Set parameters for included libraries
-            // Note: this may not be the optimal method (e.g. libraries with the same
-            // parameter name, but different
-            // values)
-            if (lib.getLibrary().getIncludes() != null) {
-                lib.getLibrary()
-                        .getIncludes()
-                        .getDef()
-                        .forEach(includeDef -> paramMap.forEach((paramKey, paramValue) -> context.getState()
-                                .setParameter(includeDef.getLocalIdentifier(), paramKey, paramValue)));
-            }
-        }
+        // setup MeasureDef
+        var measureDef = new Dstu3MeasureDefBuilder().build(measure);
 
         var actualRepo = this.repository;
         if (additionalData != null) {
             actualRepo = new FederatedRepository(
-                    this.repository, new InMemoryFhirRepository(this.repository.fhirContext(), additionalData));
+                this.repository, new InMemoryFhirRepository(this.repository.fhirContext(), additionalData));
         }
-
-        var evalType = MeasureEvalType.fromCode(reportType)
-                .orElse(
-                        subjectIds == null || subjectIds.isEmpty() || subjectIds.get(0) == null
-                                ? MeasureEvalType.POPULATION
-                                : MeasureEvalType.SUBJECT);
-
         var subjects = subjectProvider.getSubjects(actualRepo, subjectIds).collect(Collectors.toList());
-        var libraryEngine = new LibraryEngine(repository, this.measureEvaluationOptions.getEvaluationSettings());
-        Dstu3MeasureEvaluation measureEvaluator = new Dstu3MeasureEvaluation(context, measure, libraryEngine, id);
-        return measureEvaluator.evaluate(evalType, subjects, measurementPeriod, libraryEngine, id);
+        var evalType = getMeasureEvalType(reportType, subjects);
+        var context = Engines.forRepository(
+                this.repository, this.measureEvaluationOptions.getEvaluationSettings(), additionalData);
+
+        var libraryVersionIdentifier = getLibraryVersionIdentifier(measure);
+        var libraryEngine = getLibraryEngine(parameters, libraryVersionIdentifier, context);
+        // set measurement Period from CQL if operation parameters are empty
+        measureProcessorUtils.setMeasurementPeriod(measureDef, measurementPeriod, context);
+        // set offset of operation parameter measurement period
+        ZonedDateTime zonedMeasurementPeriod = MeasureProcessorUtils.getZonedTimeZoneForEval(measurementPeriod);
+        // populate results from Library $evaluate
+        var results = measureProcessorUtils.getEvaluationResults(subjectIds, measureDef, additionalData, zonedMeasurementPeriod, context, libraryEngine, libraryVersionIdentifier);
+
+        //Process Criteria Expression Results
+        measureProcessorUtils.processResults(results, measureDef, evalType,
+            measureEvaluationOptions.getApplyScoringSetMembership(), new Dstu3PopulationBasisValidator());
+
+        // Populate populationDefs that require MeasureDef results
+        // TODO JM: CLI tool is not compliant here due to requiring CQL Engine context
+        measureProcessorUtils.continuousVariableObservation(measureDef, context);
+
+        // Build Measure Report with Results
+        return new Dstu3MeasureReportBuilder().build(measure, measureDef, evalTypeToReportType(evalType), measurementPeriod, subjects);
     }
 
     protected MeasureReportType evalTypeToReportType(MeasureEvalType measureEvalType) {
@@ -150,18 +128,66 @@ public class Dstu3MeasureProcessor {
                 return MeasureReportType.SUMMARY;
             default:
                 throw new InvalidRequestException(
-                        String.format("Unsupported MeasureEvalType: %s", measureEvalType.toCode()));
+                    String.format("Unsupported MeasureEvalType: %s", measureEvalType.toCode()));
         }
     }
 
-    private Interval buildMeasurementPeriod(String periodStart, String periodEnd) {
-        // resolve the measurement period
-        return new Interval(
-                DateHelper.resolveRequestDate(periodStart, true),
-                true,
-                DateHelper.resolveRequestDate(periodEnd, false),
-                true);
+    protected LibraryEngine getLibraryEngine(Parameters parameters, VersionedIdentifier id, CqlEngine context) {
+
+        CompiledLibrary lib;
+        try {
+            lib = context.getEnvironment().getLibraryManager().resolveLibrary(id);
+        } catch (CqlIncludeException e) {
+            throw new IllegalStateException(
+                String.format(
+                    "Unable to load CQL/ELM for library: %s. Verify that the Library resource is available in your environment and has CQL/ELM content embedded.",
+                    id.getId()),
+                e);
+        }
+
+        context.getState().init(lib.getLibrary());
+
+        if (parameters != null) {
+            Map<String, Object> paramMap = resolveParameterMap(parameters);
+            context.getState().setParameters(lib.getLibrary(), paramMap);
+            // Set parameters for included libraries
+            // Note: this may not be the optimal method (e.g. libraries with the same
+            // parameter name, but different
+            // values)
+            if (lib.getLibrary().getIncludes() != null) {
+                lib.getLibrary()
+                    .getIncludes()
+                    .getDef()
+                    .forEach(includeDef -> paramMap.forEach((paramKey, paramValue) -> context.getState()
+                        .setParameter(includeDef.getLocalIdentifier(), paramKey, paramValue)));
+            }
+        }
+
+        return new LibraryEngine(repository, this.measureEvaluationOptions.getEvaluationSettings());
     }
+
+    private VersionedIdentifier getLibraryVersionIdentifier(Measure measure){
+        var reference = measure.getLibrary().get(0);
+
+        var library = this.repository.read(Library.class, reference.getReferenceElement());
+
+        return new VersionedIdentifier().withId(library.getName()).withVersion(library.getVersion());
+    }
+
+    private MeasureEvalType getMeasureEvalType(String reportType, List<String> subjectIds){
+        return MeasureEvalType.fromCode(reportType)
+            .orElse(
+                subjectIds == null || subjectIds.isEmpty() || subjectIds.get(0) == null
+                    ? MeasureEvalType.POPULATION
+                    : MeasureEvalType.SUBJECT);
+    }
+    private void checkMeasureLibrary(Measure measure){
+        if (!measure.hasLibrary()) {
+            throw new InvalidRequestException(
+                String.format("Measure %s does not have a primary library specified", measure.getUrl()));
+        }
+    }
+
 
     private Map<String, Object> resolveParameterMap(Parameters parameters) {
         Map<String, Object> parameterMap = new HashMap<>();
