+++ cqf-fhir-cr/src/main/java/org/opencds/cqf/fhir/cr/measure/r4/R4MeasureService.java
@@ -78,7 +78,7 @@ public class R4MeasureService implements R4MeasureEvaluatorSingle {
                 reportType,
                 Collections.singletonList(subjectId),
                 additionalData,
-                parameters);
+                parameters, measureEvaluationOptions.getApplyScoringSetMembership());
 
         // add ProductLine after report is generated
         measureReport = r4MeasureServiceUtils.addProductLineExtension(measureReport, productLine);
